+++ waves/tests/test_latinhypercube.py
@@ -85,12 +85,12 @@ class TestLatinHypercube:
         parameter_names = [key for key in parameter_schema.keys() if key != 'num_simulations']
         TestGenerate = LatinHypercube(parameter_schema)
         TestGenerate.generate()
-        samples_array = TestGenerate.samples
-        quantiles_array = TestGenerate.quantiles
+        samples_array = TestGenerate._samples
+        quantiles_array = TestGenerate._quantiles
         assert samples_array.shape == (parameter_schema['num_simulations'], len(parameter_names))
         assert quantiles_array.shape == (parameter_schema['num_simulations'], len(parameter_names))
         # Verify that the parameter set name creation method was called
-        assert TestGenerate.parameter_set_names == [f"parameter_set{num}" for num in range(parameter_schema['num_simulations'])]
+        assert list(TestGenerate._parameter_set_names.values()) == [f"parameter_set{num}" for num in range(parameter_schema['num_simulations'])]
         # Check that the parameter set names are correctly populated in the parameter study Xarray Dataset
         expected_set_names = [f"parameter_set{num}" for num in range(parameter_schema['num_simulations'])]
         parameter_set_names = list(TestGenerate.parameter_study[_set_coordinate_key])
@@ -102,6 +102,6 @@ class TestLatinHypercube:
                              ids=generate_input.keys())
     def test_generate_parameter_distributions(self, parameter_schema):
         TestDistributions = LatinHypercube(parameter_schema)
-        assert TestDistributions.parameter_names == list(TestDistributions.parameter_distributions.keys())
+        assert TestDistributions._parameter_names == list(TestDistributions.parameter_distributions.keys())
         # TODO: More rigorous scipy.stats object inspection to test object construction
         # https://re-git.lanl.gov/aea/python-projects/waves/-/issues/261